conda-requirements.txt
numpy
scipy
pandas
bokeh
python-dateutil
nomkl
